Move VEP tests from the source code into `tests` folder.
